backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]
mistralai
python-dotenv

pytest
